+++ sos21-api-server/CHANGELOG.md
@@ -13,6 +13,9 @@ with the HTTP API interface being the "public API".
 - Support for non-ASCII filenames ([#98](https://github.com/sohosai/sos21-backend/pull/98))
 
 ### Changed
+
+- Enable to specify project by code in file-distribution/create ([#96](https://github.com/sohosai/sos21-backend/pull/96))
+
 ### Deprecated
 ### Removed
 ### Fixed